Minor refactoring to move all runtime configs from 'config.js' into 'environment.js'.
