+++ src/webview/recipe.js
@@ -29,8 +29,7 @@ import contextMenu from './contextMenu';
 import './notifications';
 import { screenShareCss } from './screenshare';
 
-import { DEFAULT_APP_SETTINGS } from '../config';
-import { isDevMode } from '../environment';
+import { DEFAULT_APP_SETTINGS, isDevMode } from '../environment';
 
 const debug = require('debug')('Ferdi:Plugin');
